+++ packages/nuxt-module/test/fixtures/with-functions/nuxt.config.ts
@@ -1,0 +1,9 @@
+import MyModule from '../../../src/module'
+
+export default defineNuxtConfig({
+  modules: [MyModule],
+  compatibilityDate: '2025-07-11',
+  future: {
+    compatibilityVersion: 4,
+  },
+})