+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [2.4.0](https://github.com/netlify/primitives/compare/vite-plugin-v2.3.10...vite-plugin-v2.4.0) (2025-07-15)
+
+
+### Features
+
+* add `@netlify/nuxt` Nuxt module ([#344](https://github.com/netlify/primitives/issues/344)) ([efdf8df](https://github.com/netlify/primitives/commit/efdf8dfb22ffe33b557d2140fd3a3152e23cc95c))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.4.0 to 4.4.1
+    * @netlify/dev-utils bumped from ^3.2.2 to ^3.3.0
+
+## [2.3.10](https://github.com/netlify/primitives/compare/vite-plugin-v2.3.9...vite-plugin-v2.3.10) (2025-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.3.8 to 4.4.0
+
 ## [2.3.9](https://github.com/netlify/primitives/compare/vite-plugin-v2.3.8...vite-plugin-v2.3.9) (2025-07-11)